Store hooks in a separate folder
