+++ asreview/webapp/src/hooks/useKeyPress.js
@@ -1,0 +1,35 @@
+import { useEffect, useState } from 'react'
+
+const useKeyPress = (targetKey) => {
+
+  const [keyPressed, setKeyPressed] = useState(false);
+  
+  useEffect(() => {
+
+    const downHandler = ({ key }) => {
+      if (key.toLowerCase() === targetKey) {
+        setKeyPressed(true);
+      }
+    };
+
+    const upHandler = ({ key }) => {
+      if (key.toLowerCase() === targetKey) {
+        setKeyPressed(false);
+      }
+    };
+
+    // Add event listeners
+    window.addEventListener('keydown', downHandler);
+    window.addEventListener('keyup', upHandler);
+    // Remove event listeners on cleanup
+    return () => {
+      window.removeEventListener('keydown', downHandler);
+      window.removeEventListener('keyup', upHandler);
+    };
+
+  }, [targetKey]);
+
+  return keyPressed;
+};
+
+export { useKeyPress };
